api/requirements.txt
pandas
pyspark==3.1.2
mysql-connector-python
amqp==2.6.1
billiard==3.6.3.0
blinker==1.4
celery==4.4.7
click==7.1.2
Flask==1.1.2
Flask-Mail==0.9.1
Flask-SocketIO==4.3.1
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2020.1
redis==3.5.3
six==1.15.0
vine==1.3.0
Werkzeug==1.0.1
